pyproject.toml
[project]
name = "django-crawler"
version = "0.1.0"
description = "Django Web Crawler for Product Scraping"
authors = [
    {name = "hsmmi",email = "hesoyam.nomad@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pytz (>=2025.1,<2026.0)",
    "scrapy (>=2.12.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "celery (>=5.4.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false